% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_stats.R
\encoding{UTF-8}
\name{format_corr}
\alias{format_corr}
\title{Format correlation statistics}
\usage{
format_corr(
  x,
  digits = 2,
  pdigits = 3,
  pzero = FALSE,
  ci = TRUE,
  italics = TRUE,
  type = "md"
)
}
\arguments{
\item{x}{Correlation object}

\item{digits}{Number of digits after the decimal for means, confidence
intervals, and correlation coefficients}

\item{pdigits}{Number of digits after the decimal for p-values, ranging
between 1-5 (also controls cutoff for small p-values)}

\item{pzero}{Logical indicator of whether to include leading zero for
p-values}

\item{ci}{Logical indicator of whether to print 95\% confidence intervals}

\item{italics}{Logical for whether \emph{p} label should be italicized}

\item{type}{Type of formatting ("md" = markdown, "latex" = LaTeX)}
}
\value{
A character string of statistical information formatted in Markdown or LaTeX.
}
\description{
With \code{format_corr()} you can format correlation statistics generated from
\code{cor.test()} output. This detects whether the object is from a Pearson,
Kendall, or Spearman correlation and reports the appropriate correlation
label (r, τ, ρ). The default output is APA formatted, but numbers of digits,
leading zeros, the presence of confidence intervals, and italics are all
customizable.
}
\examples{
# Prepare data
mtcars_corr <- cor.test(mtcars$mpg, mtcars$disp)
# Print statistics
format_corr(mtcars_corr)
# Change digits
format_corr(mtcars_corr, digits = 3)
# Change cutoff digits for p-value
format_corr(mtcars_corr, pdigits = 2)
# Add leading zero to p-value and don't print confidence intervals
format_corr(mtcars_corr, pzero = TRUE, ci = FALSE)
}
\seealso{
Other functions for printing statistical objects: 
\code{\link{format_bf}()},
\code{\link{format_ttest}()}
}
\concept{functions for printing statistical objects}
