% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_numbers.R
\name{format_mean}
\alias{format_mean}
\title{Calculate and format mean and error}
\usage{
format_mean(
  x = NULL,
  error = "ci",
  values = NULL,
  digits = 1,
  meanlabel = "abbr",
  italics = TRUE,
  subscript = NULL,
  units = NULL,
  display = "limits",
  cilevel = 0.95,
  errorlabel = TRUE,
  type = "md"
)
}
\arguments{
\item{x}{Numeric vector of data to calculate mean and error}

\item{error}{Character vector specifying error type ("ci" = confidence
interval, "se" = standard error of the mean, "sd" = standard deviation)}

\item{values}{Numeric vector of mean and interval or mean and lower and upper
limits}

\item{digits}{Number of digits after the decimal for means and error}

\item{meanlabel}{Formatting for mean label ("abbr" = M, "word" = Mean,
"none" = no label)}

\item{italics}{Logical for whether mean label should be italicized}

\item{subscript}{Character string to include as subscript with mean label}

\item{units}{Character string that gives units to include after mean value}

\item{display}{Character vector specifying how to display error ("limits" =
[lower limit, upper limit], "pm" = ±interval, "par" = (interval), "none" =
do not display error)}

\item{cilevel}{Numeric scalar from 0-1 defining confidence level
(defaults to 0.95)}

\item{errorlabel}{Logical for whether error label (e.g., 95\% CI) should be
included}

\item{type}{Type of formatting ("md" = markdown, "latex" = LaTeX)}
}
\value{
A character string of mean and error formatted in Markdown or LaTeX.
To return only the mean (no error), set \code{display = "none"}.
}
\description{
Calculate and format mean and error
}
\examples{
# Print mean and 95\% confidence limits for fuel efficiency
format_mean(mtcars$mpg)
# Print mean and standard error of the mean as interval
format_mean(mtcars$mpg, error = "se", display = "pm")
# Print mean and standard deviation as interval
format_mean(mtcars$mpg, error = "sd", display = "par")
# Print mean and 90\% confidence limits with units
format_mean(mtcars$mpg, units = "cm", cilevel = 0.9)
# Print two-digit mean with subscript in LaTeX
format_mean(mtcars$mpg, digits = 2, subscript = "control", display = "none", type = "latex")
}
