% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_numbers.R
\name{format_meanci}
\alias{format_meanci}
\title{Calculate and format mean and confidence interval}
\usage{
format_meanci(
  x = NULL,
  values = NULL,
  level = 0.95,
  digits = 1,
  italics = TRUE,
  mean = "abbr",
  type = "md"
)
}
\arguments{
\item{x}{Numeric vector of data to calculate mean and confidence interval}

\item{values}{Numeric vector of mean, lower CI, and upper CI to format}

\item{level}{Numeric scalar from 0-1 defining width of confidence interval
(defaults to 0.95)}

\item{digits}{Number of digits after the decimal for means and confidence
intervals}

\item{italics}{Logical for whether \emph{p} label should be italicized}

\item{mean}{Formatting for mean label ("abbr" = M, "word" = Mean)}

\item{type}{Type of formatting ("md" = markdown, "latex" = LaTeX)}
}
\description{
Calculate and format mean and confidence interval
}
