% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{format_num}
\alias{format_num}
\title{Format numbers}
\usage{
format_num(x, digits)
}
\arguments{
\item{x}{Number}

\item{digits}{Number of digits after the decimal}
}
\value{
A character string formatting the number
}
\description{
Format numbers
}
\examples{
format_num(pi, digits = 2)
format_num(pi, digits = 4)
}
