% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_numbers.R
\name{format_p}
\alias{format_p}
\title{Format p-values}
\usage{
format_p(x, pdigits = 3, pzero = FALSE, italics = TRUE, type = "md")
}
\arguments{
\item{x}{Number representing p-value}

\item{pdigits}{Number of digits after the decimal for p-values (also controls
cutoff for small p-values)}

\item{pzero}{Logical indicator of whether to include leading zero for
p-values}

\item{italics}{Logical for whether \emph{p} label should be italicized or not}

\item{type}{Type of formatting ("md" = markdown, "latex" = LaTeX)}
}
\value{
A character string that includes \emph{p} and then the p-value formatted in
Markdown. If p-value is below \code{pdigits} cutoff, \verb{_p_ < <cutoff>} is used.
}
\description{
Format p-values
}
\examples{
format_p(0.001)
# Round digits for p-values greater than cutoff
format_p(0.111, pdigits = 2)
# Default cutoff is p < 0.001
format_p(0.0001)
# Set cutoff with pdigits
format_p(0.0001, pdigits = 2)
# Include leading zero
format_p(0.001, pzero = TRUE)
}
