% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_numbers.R
\name{format_scientific}
\alias{format_scientific}
\title{Format numbers in scientific notation}
\usage{
format_scientific(x, digits = 1, type = "md")
}
\arguments{
\item{x}{Number}

\item{digits}{Number of digits after the decimal}

\item{type}{Type of formatting ("md" = markdown, "latex" = LaTeX)}
}
\value{
A character string of a number in scientific notation formatted in Markdown
or LaTeX.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

All functions in this package are deprecated. Please use the
\href{https://github.com/JeffreyRStevens/cocoon}{\code{{cocoon}}} package instead.
}
\examples{
format_scientific(1111)
# Control number of digits after decimal with digits
format_scientific(1111, digits = 3)
}
\keyword{internal}
