% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatstats-package.R
\docType{package}
\name{formatstats-package}
\alias{formatstats}
\alias{formatstats-package}
\title{formatstats: Format and Print Statistical Output}
\description{
Provides functions that format statistical output in a way that can be inserted into R Markdown documents. This is analogous to the `apa_print()` functions in the papaja package but prints Markdown or LaTeX syntax.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/JeffreyRStevens/formatstats}
  \item \url{https://jeffreyrstevens.github.io/formatstats/}
  \item Report bugs at \url{https://github.com/JeffreyRStevens/formatstats/issues}
}

}
\author{
\strong{Maintainer}: Jeffrey R. Stevens \email{jeffrey.r.stevens@protonmail.com} (\href{https://orcid.org/0000-0003-2375-1360}{ORCID}) [copyright holder]

}
\keyword{internal}
