% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_stats.R
\name{format_corr}
\alias{format_corr}
\title{Format correlation statistics}
\usage{
format_corr(
  x,
  digits = 2,
  pdigits = 3,
  pzero = FALSE,
  ci = TRUE,
  italics = TRUE,
  dfs = "par",
  type = "md"
)
}
\arguments{
\item{x}{Correlation object}

\item{digits}{Number of digits after the decimal for means, correlation
coefficient, confidence intervals, and degrees of freedom}

\item{pdigits}{Number of digits after the decimal for p-values (also controls
cutoff for small p-values)}

\item{pzero}{Logical indicator of whether to include leading zero for
p-values}

\item{ci}{Logical indicator of whether to print 95\% confidence intervals}

\item{italics}{Logical for whether \emph{p} label should be italicized or not}

\item{dfs}{Formatting for degrees of freedom ("par" = parenthetical,
"sub" = subscript, "none" = do not print degrees of freedom)}

\item{type}{Type of formatting ("md" = markdown, "latex" = LaTeX)}
}
\value{
A character string of statistical information formatted in Markdown.
}
\description{
Format correlation statistics
}
\examples{
# Prepare data
beavers <- merge(beaver1, beaver2, by = "time")
beavers <- beavers[, c("time", "temp.x", "temp.y")]
beavers_corr <- cor.test(beavers$temp.x, beavers$temp.y)
# Print statistics
format_corr(beavers_corr)
# Change digits
format_corr(beavers_corr, digits = 3)
# Change cutoff digits for p-value
format_corr(beavers_corr, pdigits = 2)
# Add leading zero to p-value and don't print confidence intervals
format_corr(beavers_corr, pzero = TRUE, ci = FALSE)
}
\seealso{
Other functions for printing statistics: 
\code{\link{format_ttest}()}
}
\concept{functions for printing statistics}
