% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_numbers.R
\name{format_num}
\alias{format_num}
\title{Format numbers}
\usage{
format_num(x, digits = 1, pzero = TRUE)
}
\arguments{
\item{x}{Number}

\item{digits}{Number of digits after the decimal}

\item{pzero}{Logical indicator of whether to include leading zero numbers
less than 1}
}
\value{
A character string formatting the number with specified number of digits
after the decimal.
}
\description{
Format numbers
}
\examples{
format_num(pi, digits = 2)
format_num(pi, digits = 4)
}
