% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_stats.R
\name{format_p}
\alias{format_p}
\title{Format p-values}
\usage{
format_p(
  x,
  digits = 3,
  pzero = FALSE,
  label = "p",
  italics = TRUE,
  type = "md"
)
}
\arguments{
\item{x}{Number representing p-value}

\item{digits}{Number of digits after the decimal for p-values, ranging
between 1-5 (also controls cutoff for small p-values)}

\item{pzero}{Logical indicator of whether to include leading zero for
p-values}

\item{label}{Character string for label before p value. Default is p.
Set \code{label = ""} to return just the formatted p value with no
label or operator (\code{=}, \code{<}, \code{>})}

\item{italics}{Logical for whether label should be italicized (\emph{p})}

\item{type}{Type of formatting ("md" = markdown, "latex" = LaTeX)}
}
\value{
A character string that includes \emph{p} and then the p-value formatted in
Markdown or LaTeX. If p-value is below \code{digits} cutoff, \code{p < cutoff} is
used.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

All functions in this package are deprecated. Please use the
\href{https://github.com/JeffreyRStevens/cocoon}{\code{{cocoon}}} package instead.

\code{format_p()} inputs numeric vectors of p-values. Cutoffs can be set that
format the values as less than the cutoffs (e.g., p < 0.001). The default
output is APA formatted, but numbers of digits, cutoffs, leading zeros, and
italics are all customizable.
}
\examples{
format_p(0.001)
# Round digits for p-values greater than cutoff
format_p(0.111, digits = 2)
# Default cutoff is p < 0.001
format_p(0.0001)
# Set cutoff with digits
format_p(0.0001, digits = 2)
# Include leading zero
format_p(0.001, pzero = TRUE)
# Return only Bayes factor value (no label)
format_p(0.001, label = "")
}
\keyword{internal}
