% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_stats.R
\name{format_ttest}
\alias{format_ttest}
\title{Format t-test statistics}
\usage{
format_ttest(
  x,
  digits = 1,
  pdigits = 3,
  pzero = FALSE,
  full = TRUE,
  italics = TRUE,
  dfs = "par",
  mean = "abbr",
  type = "md"
)
}
\arguments{
\item{x}{t-test object}

\item{digits}{Number of digits after the decimal for means, confidence
intervals, and t-statistics}

\item{pdigits}{Number of digits after the decimal for p-values, ranging
between 1-5 (also controls cutoff for small p-values)}

\item{pzero}{Logical indicator of whether to include leading zero for
p-values}

\item{full}{Logical indicator of whether to include means and confidence
intervals or just t-statistic and p-value}

\item{italics}{Logical for whether \emph{p} label should be italicized}

\item{dfs}{Formatting for degrees of freedom ("par" = parenthetical,
"sub" = subscript, "none" = do not print degrees of freedom)}

\item{mean}{Formatting for mean label ("abbr" = M, "word" = Mean)}

\item{type}{Type of formatting ("md" = markdown, "latex" = LaTeX)}
}
\value{
A character string of statistical information formatted in Markdown or LaTeX.
}
\description{
Format t-test statistics
}
\examples{
# Prepare data
mtcars_tt <- t.test(formula = mtcars$mpg ~ mtcars$vs)
# Print statistics
format_ttest(mtcars_tt)
# Change digits
format_ttest(mtcars_tt, digits = 2)
# Change cutoff digits for p-value
format_ttest(mtcars_tt, pdigits = 2)
# Add leading zero to p-value and don't print confidence intervals
format_ttest(mtcars_tt, pzero = TRUE, full = FALSE)
}
\seealso{
Other functions for printing statistics: 
\code{\link{format_corr}()}
}
\concept{functions for printing statistics}
